refactor(host-application): reorganize build orchestrator lifecycle and runtime setup
