+++ apps/desktop/src-tauri/crates/host-application/src/app_service/build_orchestrator/build_cleanup.rs
@@ -1,5 +1,6 @@
 use super::super::{
-    emit_event, run_parsed_hook_command_allow_failure, AppService, RunEmitter, RunProcess,
+    emit_event, run_parsed_hook_command_allow_failure, terminate_child_process, AppService,
+    RunEmitter, RunProcess,
 };
 use super::CleanupMode;
 use anyhow::{anyhow, Result};
@@ -29,7 +30,7 @@ impl AppService {
     ) -> Result<bool> {
         let mut run = self.take_run_for_cleanup(run_id)?;
 
-        super::super::terminate_child_process(&mut run.child);
+        terminate_child_process(&mut run.child);
 
         match mode {
             CleanupMode::Failure => self.cleanup_failed_run(run_id, &mut run, &emitter),
